+++ Specs/Tools/PequenoNetworking/PequenoNetworkingConstantsSpec.swift
@@ -1,0 +1,15 @@
+import Quick
+import Moocher
+@testable import Utensils
+
+final class PequenoNetworkingConstantsSpec: QuickSpec {
+    override func spec() {
+        describe("PequenoNetworkingConstants") {
+            it("has all required constants") {
+                expect(PequenoNetworkingConstants.BaseURLKey).to.equal("PequenoNetworkingBaseURLKey")
+                
+                expect(PequenoNetworkingConstants.HeadersKey).to.equal("PequenoNetworkingHeadersKey")
+            }
+        }
+    }
+}
